refactor: move error to it's own file
